requirements.txt
joblib~=1.0.0
numpy~=1.21.5
opencv-contrib-python<=3.4.2.17
Pillow~=9.0.0
scikit-learn~=1.0.2
scipy~=1.7.3
tqdm~=4.62.3
tk~=0.1.0
